+++ packages/website/src/pages/documentation/index.page.server.tsx
@@ -1,6 +1,9 @@
 import type { OnBeforeRender } from "@src/renderer/types.js"
 import type { PageProps } from "./index.page.jsx"
-import { tableOfContents, FrontmatterSchema } from "../../../../../documentation/tableOfContents.js"
+import {
+	FrontmatterSchema,
+	tableOfContentsPromise,
+} from "../../../../../documentation/tableOfContents.js"
 import { parseMarkdown } from "@src/services/markdown/index.js"
 import { RenderErrorPage } from "vite-plugin-ssr/RenderErrorPage"
 
@@ -14,30 +17,11 @@ export type ProcessedTableOfContents = Record<
 	Omit<(typeof index)[keyof typeof index], "html">[]
 >
 
-/**
- * The index of documentation markdown files. The href's acts as id.
- *
- * @example
- * 	{
- * 		"/documentation/intro": document,
- * 	}
- */
-const index: Record<string, Awaited<ReturnType<typeof parseMarkdown>>> = {}
-/**
- * the table of contents without the html for each document
- * saving bandwith and speeding up the site)
- */
-const processedTableOfContents: ProcessedTableOfContents = {}
-
-await generateIndexAndTableOfContents()
-
 // should only run server side
 export const onBeforeRender: OnBeforeRender<PageProps> = async (pageContext) => {
-	let headings = []
 	// dirty way to get reload of markdown (not hot reload though)
 	if (import.meta.env.DEV) {
 		await generateIndexAndTableOfContents()
-		headings = await generateHeadings(pageContext.urlPathname)
 	}
 	if (!Object.keys(index).includes(pageContext.urlPathname)) {
 		throw RenderErrorPage({ pageContext: { is404: true } })
@@ -47,50 +31,48 @@ export const onBeforeRender: OnBeforeRender<PageProps> = async (pageContext) => 
 			pageProps: {
 				markdown: index[pageContext.urlPathname]!,
 				processedTableOfContents: processedTableOfContents,
-				headings: headings,
 			},
 		},
 	}
 }
 
 /**
- * Generates the index and table of contents.
+ * Generates the index and table of contents using async/await.
  */
 async function generateIndexAndTableOfContents() {
-	for (const [category, documents] of Object.entries(tableOfContents)) {
-		const frontmatters: { frontmatter: any }[] = []
-		for (const document of documents) {
+	try {
+		const tableOfContents = await tableOfContentsPromise
+
+		const processedTableOfContents: { [key: string]: { frontmatter: any }[] } = {}
+
+		/**
+		 * The index of documentation markdown files. The href's acts as id.
+		 *
+		 * @example
+		 * 	{
+		 * 		"/documentation/intro": document,
+		 * 	}
+		 */
+		const index: { [key: string]: any } = {}
+
+		for (const { category, content } of tableOfContents) {
 			const markdown = parseMarkdown({
-				text: document,
+				text: content,
 				FrontmatterSchema,
 			})
-			// not pushing to processedTableOfContents directly in case
-			// the category is undefined so far
-			frontmatters.push({ frontmatter: markdown.frontmatter })
+			if (!processedTableOfContents[category]) {
+				processedTableOfContents[category] = []
+			}
+			processedTableOfContents[category]?.push({ frontmatter: markdown.frontmatter })
 			index[markdown.frontmatter.href] = markdown
 		}
-		processedTableOfContents[category] = frontmatters
+
+		return { processedTableOfContents, index }
+	} catch (error) {
+		console.error("Error generating index and table of contents:", error)
+		throw error
 	}
 }
 
-/**
- * Generates the headings
- */
-async function generateHeadings(urlPathname: string) {
-	const headings: PageProps["headings"] = []
-
-	const markdown = index[urlPathname]
-	if (markdown && markdown.renderableTree) {
-		for (const heading of markdown.renderableTree.children) {
-			if (heading.name === "Heading") {
-				if (heading.children[0].name) {
-					headings.push(heading.children[0])
-				} else {
-					headings.push(heading)
-				}
-			}
-		}
-	}
-
-	return headings
-}
+// Call the function to generate the index and table of contents.
+const { processedTableOfContents, index } = await generateIndexAndTableOfContents()