+++ packages/website/src/pages/index/sections/04-pricing/index.tsx
@@ -6,7 +6,7 @@ const Pricing = () => {
 
 	return (
 		<SectionLayout type="white" showLines={true}>
-			<div class="flex justify-between px-10 py-16 sm:py-20 lg:py-32 gap-16 lg:gap-0 flex-col lg:flex-row">
+			<div class="flex justify-between px-6 md:px-10 py-16 sm:py-20 lg:py-32 gap-16 lg:gap-0 flex-col lg:flex-row">
 				<div class="w-full lg:w-[calc((100%_-_40px)_/_2)] flex flex-col gap-2 lg:gap-4 items-center lg:items-start">
 					<p class="text-sm text-center lg:text-start text-primary bg-primary/10 h-7 flex items-center px-4 rounded-full w-fit tracking-relaxed">
 						{t("landing.pricing.caption")}
@@ -14,7 +14,7 @@ const Pricing = () => {
 					<h2 class="text-3xl text-center lg:text-start  font-semibold text-on-background leading-tight md:leading-relaxed tracking-tight">
 						{t("landing.pricing.title")}
 					</h2>
-					<p class="text-base text-center lg:text-start md:w-[80%] text-outline-variant sm:leading-7">
+					<p class="text-base text-center lg:text-start md:w-[80%] text-outline-variant sm:leading-7 pb-16">
 						{t("landing.pricing.description")}
 					</p>
 				</div>